Secure user & transaction modules with JWT authentication

+++ backend/server/controllers/transactionCtrl.js
@@ -1,11 +1,8 @@
 const Transaction = require("../models/transaction")
 
-module.exports.getAllTransaction = (req, res, next) => {
-  Transaction.find()
+module.exports.getAllTransaction = async (req, res, next) => {
+  Transaction.find({ userId: req.auth.userId })
     .then((transactions) => {
-      if (!transactions) {
-        res.status(400).json({ message: "Invalid request" })
-      }
       res.status(200).json(transactions)
     })
     .catch((error) => {
@@ -15,12 +12,13 @@ module.exports.getAllTransaction = (req, res, next) => {
 module.exports.getOneTransaction = (req, res, next) => {
   Transaction.findOne({
     _id: req.params.id,
+    userId: req.auth.userId,
   })
     .then((transaction) => {
       if (!transaction) {
-        res.status(404).json({ message: "Transaction not found" })
+        return res.status(404).json({ message: "Transaction not found" })
       }
-      res.status(200).json(transaction)
+      return res.status(200).json(transaction)
     })
     .catch((error) => {
       res.status(500).json({ error: "Erreur serveur", details: error.message })
@@ -37,6 +35,7 @@ module.exports.createTransaction = (req, res, next) => {
 
   const transaction = new Transaction({
     ...req.body,
+    userId: req.auth.userId,
     date: dateOnly,
   })
   transaction
@@ -50,7 +49,10 @@ module.exports.createTransaction = (req, res, next) => {
 }
 module.exports.updateOneTransaction = (req, res, next) => {
   const transaction = new Transaction({ ...req.body })
-  Transaction.updateOne({ _id: req.params.id }, transaction)
+  Transaction.updateOne(
+    { _id: req.params.id, userId: req.auth.userId },
+    transaction
+  )
     .then(() => {
       res.status(201).json({
         message: "Transaction updated Successfully!",
@@ -61,15 +63,17 @@ module.exports.updateOneTransaction = (req, res, next) => {
     })
 }
 module.exports.deleteOneTransaction = (req, res, next) => {
-  Transaction.deleteOne({ _id: req.params.id })
-    .then(() => {
-      res.status(200).json({
-        message: "Transaction deleted !",
-      })
+  Transaction.deleteOne({ _id: req.params.id, userId: req.auth.userId })
+    .then((result) => {
+      if (result.deletedCount === 0) {
+        return res
+          .status(404)
+          .json({ message: "Transaction not found or not yours" })
+      } else {
+        return res.status(200).json({ message: "Transaction deleted!" })
+      }
     })
     .catch((error) => {
-      res.status(404).json({
-        error,
-      })
+      res.status(500).json({ error: "Erreur serveur", details: error.message })
     })
 }
